change signature to int64_t; rm streaming predict and fit_predict

+++ cpp/src/cluster/kmeans_fit_double.cu
@@ -66,13 +66,13 @@ void fit(raft::resources const& handle,
 
 void fit(raft::resources const& handle,
          const cuvs::cluster::kmeans::params& params,
-         raft::host_matrix_view<const double, int64_t> X,
-         std::optional<raft::host_vector_view<const double, int64_t>> sample_weight,
-         raft::device_matrix_view<double, int64_t> centroids,
+         raft::host_matrix_view<const double, int> X,
+         std::optional<raft::host_vector_view<const double, int>> sample_weight,
+         raft::device_matrix_view<double, int> centroids,
          raft::host_scalar_view<double> inertia,
-         raft::host_scalar_view<int64_t> n_iter)
+         raft::host_scalar_view<int> n_iter)
 {
-  cuvs::cluster::kmeans::detail::fit<double, int64_t>(
+  cuvs::cluster::kmeans::detail::fit<double, int>(
     handle, params, X, sample_weight, centroids, inertia, n_iter);
 }
